MVP Met - Add useLocalStorage & useDarkMode js file Because: - Built a custom hook that let's you save data to localStorage - Built a second custom hook that sets the dark-mode class on the body element - Composed my two new hooks together to be able to set and persist my user's dark mode preference in my app
This commit:
- Step 1: Created a new directory called hooks, and a new file in it called useLocalStorage
Built a function called useLocalStorage. To set sth. to localStorage, I added a key (string) and a value. To retrieve
sth. from localStorage, I need the key. To update sth. in localStorage, I use the same method as adding sth. new, and
it will replace the old key/value pair in localStorage. Added key and initialValue as parameters to the hook.
- Step 2: Added a new file called useDarkMode inside the hooks directory.
Built a function called useDarkMode
Imported useLocalStorage
- Step 3: Used the hook in a component.